refactor: support most types for opacity
