+++ extension/src/shared/layerContainers/hooks/useEvaluateGeneralAppearance.ts
@@ -1,21 +1,8 @@
 import { merge } from "lodash-es";
 import { useMemo } from "react";
 
-import { isNotNullish } from "../../../prototypes/type-helpers";
-import { COLOR_MAPS } from "../../constants";
-import {
-  color,
-  conditionWithOperation,
-  defaultConditionalNumber,
-  number,
-  rgba,
-  string,
-  variable,
-} from "../../helpers";
 import { useOptionalAtomValue } from "../../hooks";
 import { GeneralAppearances } from "../../reearth/layers";
-import { ExpressionContainer } from "../../reearth/types/expression";
-import { Component } from "../../types/fieldComponents";
 import {
   TILESET_CLIPPING,
   TILESET_FILL_COLOR_CONDITION_FIELD,
@@ -59,394 +46,22 @@ import {
 } from "../../types/fieldComponents/polyline";
 import { ComponentAtom } from "../../view-layers/component";
 import { useFindComponent } from "../../view-layers/hooks";
+import {
+  DEFAULT_COLOR,
+  makeConditionalExpression,
+  makeConditionalImageColorExpression,
+  makeConditionalImageExpression,
+  makeGradientExpression,
+  makeLabelTextExpression,
+  makeSimpleColorValue,
+  makeSimpleColorWithOpacity,
+  makeSimpleValueForStrokeColor,
+  makeStrokeColorConditionalExpression,
+  makeVisibilityConditionExpression,
+  makeVisibilityFilterExpression,
+} from "../utils/value";
 
 import { useClippingBox } from "./useClippingBox";
-
-const DEFAULT_COLOR = "#ffffff";
-
-export const makeSimpleValue = (
-  comp:
-    | Component<
-        | typeof POINT_FILL_COLOR_VALUE_FIELD
-        | typeof POLYLINE_FILL_COLOR_VALUE_FIELD
-        | typeof POLYGON_FILL_COLOR_VALUE_FIELD
-      >
-    | undefined,
-): string | undefined => {
-  if (!comp) return;
-
-  switch (comp.type) {
-    // Point
-    case POINT_FILL_COLOR_VALUE_FIELD:
-      return comp.value?.color || comp.preset?.defaultValue;
-    // Polyline
-    case POLYLINE_FILL_COLOR_VALUE_FIELD:
-      return comp.value?.color || comp.preset?.defaultValue;
-    // Polygon
-    case POLYGON_FILL_COLOR_VALUE_FIELD:
-      return comp.value?.color || comp.preset?.defaultValue;
-    default:
-      return;
-  }
-};
-
-export const makeSimpleValueForStrokeColor = (
-  comp: Component<typeof POLYGON_FILL_COLOR_VALUE_FIELD> | undefined,
-): string | undefined => {
-  if (!comp) return;
-  let strokeColor;
-  let fillColor;
-  switch (comp.type) {
-    // Polygon
-    case POLYGON_FILL_COLOR_VALUE_FIELD:
-      strokeColor = comp.value?.strokeColor || comp.preset?.strokeValue;
-      fillColor = comp.value?.color || comp.preset?.defaultValue;
-      return strokeColor === "" ? fillColor : strokeColor;
-    default:
-      return;
-  }
-};
-
-export const makeConditionalExpression = (
-  comp:
-    | Component<
-        | typeof POINT_FILL_COLOR_CONDITION_FIELD
-        | typeof TILESET_FILL_COLOR_CONDITION_FIELD
-        | typeof POLYLINE_FILL_COLOR_CONDITION_FIELD
-        | typeof POLYGON_FILL_COLOR_CONDITION_FIELD
-      >
-    | undefined,
-  opacity = 1,
-): ExpressionContainer | undefined => {
-  if (!comp) return;
-
-  const currentRuleId = comp.value?.useDefault
-    ? comp.value?.currentRuleId ?? comp.preset?.rules?.[0].id
-    : comp.value?.currentRuleId;
-
-  return {
-    expression: {
-      conditions: [
-        ...(
-          comp.preset?.rules?.flatMap(rule => {
-            if (rule.id !== currentRuleId) return;
-            const overriddenRules = comp.value?.overrideRules.filter(r => r.ruleId === rule.id);
-            return rule.conditions?.map(cond => {
-              const overriddenCondition = overriddenRules?.find(r => r.conditionId === cond.id);
-              const colorValue = overriddenCondition?.color || cond.color;
-              if (!rule.propertyName || !cond.value || !colorValue) return;
-              const stringCondition = conditionWithOperation(
-                variable(rule.propertyName),
-                string(cond.value),
-                cond.operation,
-              );
-              const numberCondition = !isNaN(Number(cond.value))
-                ? conditionWithOperation(
-                    defaultConditionalNumber(rule.propertyName),
-                    number(Number(cond.value)),
-                    cond.operation,
-                  )
-                : undefined;
-              return rule.propertyName && cond.value && colorValue
-                ? ([
-                    numberCondition ? numberCondition : stringCondition,
-                    color(colorValue, opacity),
-                  ] as [string, string])
-                : undefined;
-            });
-          }) ?? []
-        ).filter(isNotNullish),
-        ["true", color(DEFAULT_COLOR, opacity)],
-      ],
-    },
-  };
-};
-
-export const makeStrokeColorConditionalExpression = (
-  comp: Component<typeof POLYGON_FILL_COLOR_CONDITION_FIELD> | undefined,
-): ExpressionContainer | undefined => {
-  if (!comp) return;
-
-  const currentRuleId = comp.value?.useDefault
-    ? comp.value?.currentRuleId ?? comp.preset?.rules?.[0].id
-    : comp.value?.currentRuleId;
-
-  return {
-    expression: {
-      conditions: [
-        ...(
-          comp.preset?.rules?.flatMap(rule => {
-            if (rule.id !== currentRuleId) return;
-            const overriddenRules = comp.value?.overrideRules.filter(r => r.ruleId === rule.id);
-            return rule.conditions?.map(cond => {
-              const overriddenCondition = overriddenRules?.find(r => r.conditionId === cond.id);
-              const colorValue = overriddenCondition?.strokeColor || cond.strokeColor;
-              if (!rule.propertyName || !cond.value || !colorValue) return;
-              const stringCondition = conditionWithOperation(
-                variable(rule.propertyName),
-                string(cond.value),
-                cond.operation,
-              );
-              const numberCondition = !isNaN(Number(cond.value))
-                ? conditionWithOperation(
-                    defaultConditionalNumber(rule.propertyName),
-                    number(Number(cond.value)),
-                    cond.operation,
-                  )
-                : undefined;
-              return rule.propertyName && cond.value && colorValue
-                ? ([numberCondition ? numberCondition : stringCondition, color(colorValue, 1)] as [
-                    string,
-                    string,
-                  ])
-                : undefined;
-            });
-          }) ?? []
-        ).filter(isNotNullish),
-        ["true", color(DEFAULT_COLOR, 0)],
-      ],
-    },
-  };
-};
-
-export const makeGradientExpression = (
-  comp:
-    | Component<typeof POINT_FILL_COLOR_GRADIENT_FIELD | typeof TILESET_FILL_COLOR_GRADIENT_FIELD>
-    | undefined,
-  opacity = 1,
-): ExpressionContainer | undefined => {
-  if (!comp) return;
-
-  const preset = comp.preset;
-  const value = comp.value;
-  const currentRuleId = comp.value?.useDefault
-    ? comp.value?.currentRuleId ?? comp.preset?.rules?.[0].id
-    : comp.value?.currentRuleId;
-  const rule = preset?.rules?.find(r => r.id === currentRuleId);
-
-  const conditions: [string, string][] = [["true", color(DEFAULT_COLOR, opacity)]];
-
-  const [minValue, maxValue] = [
-    value?.currentMin ?? rule?.min ?? 0,
-    value?.currentMax ?? rule?.max ?? 0,
-  ];
-  if (minValue === maxValue) {
-    return {
-      expression: { conditions },
-    };
-  }
-
-  const colorMap = COLOR_MAPS.find(
-    c => c.name === (value?.currentColorMapName ?? rule?.colorMapName),
-  );
-  const colorProperty = rule?.propertyName;
-
-  if (!colorMap || !colorProperty) return { expression: { conditions } };
-
-  const distance = 5;
-  for (let i = minValue; i <= maxValue; i += distance) {
-    const color = colorMap.linear((i - minValue) / (maxValue - minValue));
-    conditions.unshift([
-      `${defaultConditionalNumber(colorProperty, minValue - 1)} >= ${number(i)}`,
-      rgba({ r: color[0] * 255, g: color[1] * 255, b: color[2] * 255, a: opacity }),
-    ]);
-  }
-
-  return {
-    expression: {
-      conditions,
-    },
-  };
-};
-
-const makeSimpleColorWithOpacity = (
-  comp: Component<typeof OPACITY_FIELD> | undefined,
-  defaultColor = DEFAULT_COLOR,
-) => {
-  if (!comp) return;
-  return {
-    expression: {
-      conditions: [["true", color(defaultColor, comp.value ?? 1)]],
-    },
-  };
-};
-
-const makeVisibilityConditionExpression = (
-  comp:
-    | Component<
-        | typeof POINT_VISIBILITY_CONDITION_FIELD
-        | typeof POLYLINE_VISIBILITY_CONDITION_FIELD
-        | typeof POLYGON_VISIBILITY_CONDITION_FIELD
-      >
-    | undefined,
-): ExpressionContainer | undefined => {
-  const conditions = comp?.preset?.conditions;
-
-  if (!conditions) return;
-
-  return {
-    expression: {
-      conditions: conditions.reduce(
-        (res, cond) => {
-          const isNumber = !isNaN(Number(cond.value));
-          if (!cond.operation || !cond.value || !cond.propertyName) return res;
-          res.unshift([
-            isNumber
-              ? conditionWithOperation(
-                  defaultConditionalNumber(cond.propertyName),
-                  cond.value,
-                  cond.operation,
-                )
-              : conditionWithOperation(
-                  variable(cond.propertyName),
-                  string(cond.value),
-                  cond.operation,
-                ),
-            cond.show ? "true" : "false",
-          ]);
-          return res;
-        },
-        [["true", "false"]],
-      ),
-    },
-  };
-};
-
-const makeVisibilityFilterExpression = (
-  comp:
-    | Component<
-        | typeof POINT_VISIBILITY_FILTER_FIELD
-        | typeof POLYLINE_VISIBILITY_FILTER_FIELD
-        | typeof POLYGON_VISIBILITY_FILTER_FIELD
-      >
-    | undefined,
-): ExpressionContainer | undefined => {
-  const rule =
-    comp?.preset?.rules?.find(rule => rule.id === comp.value) ?? comp?.preset?.rules?.[0];
-  const property = rule?.propertyName;
-
-  if (!rule?.conditions || !property) return;
-
-  return {
-    expression: {
-      conditions: rule.conditions.reduce(
-        (res, cond) => {
-          const isNumber = !isNaN(Number(cond.value));
-          if (!cond.operation || !cond.value) return res;
-          res.unshift([
-            isNumber
-              ? conditionWithOperation(
-                  defaultConditionalNumber(property),
-                  cond.value,
-                  cond.operation,
-                )
-              : conditionWithOperation(variable(property), string(cond.value), cond.operation),
-            "true",
-          ]);
-          return res;
-        },
-        [["true", "false"]],
-      ),
-    },
-  };
-};
-
-export const makeConditionalImageExpression = (
-  comp: Component<typeof POINT_USE_IMAGE_CONDITION_FIELD> | undefined,
-): ExpressionContainer | undefined => {
-  if (!comp) return;
-  const currentRuleId = comp.value?.currentRuleId ?? comp.preset?.rules?.[0].id;
-  return {
-    expression: {
-      conditions: [
-        ...(
-          comp.preset?.rules?.flatMap(rule => {
-            if (rule.id !== currentRuleId) return;
-            const overriddenRules = comp.value?.overrideRules.filter(r => r.ruleId === rule.id);
-            return rule.conditions?.map(cond => {
-              const overriddenCondition = overriddenRules?.find(r => r.conditionId === cond.id);
-              const imageURLValue = overriddenCondition?.imageURL || cond.imageURL;
-              if (!rule.propertyName || !cond.value || !imageURLValue) return;
-              const stringCondition = conditionWithOperation(
-                variable(rule.propertyName),
-                string(cond.value),
-                cond.operation,
-              );
-              const numberCondition = !isNaN(Number(cond.value))
-                ? conditionWithOperation(
-                    defaultConditionalNumber(rule.propertyName),
-                    number(Number(cond.value)),
-                    cond.operation,
-                  )
-                : undefined;
-              return rule.propertyName && cond.value && imageURLValue
-                ? ([
-                    numberCondition ? `${numberCondition} || ${stringCondition}` : stringCondition,
-                    `"${imageURLValue}"`,
-                  ] as [string, string])
-                : undefined;
-            });
-          }) ?? []
-        ).filter(isNotNullish),
-      ],
-    },
-  };
-};
-
-export const makeConditionalImageColorExpression = (
-  comp: Component<typeof POINT_USE_IMAGE_CONDITION_FIELD> | undefined,
-): ExpressionContainer | undefined => {
-  if (!comp) return;
-  const currentRuleId = comp.value?.currentRuleId ?? comp.preset?.rules?.[0].id;
-  return {
-    expression: {
-      conditions: [
-        ...(
-          comp.preset?.rules?.flatMap(rule => {
-            if (rule.id !== currentRuleId) return;
-            const overriddenRules = comp.value?.overrideRules.filter(r => r.ruleId === rule.id);
-            return rule.conditions?.map(cond => {
-              const overriddenCondition = overriddenRules?.find(r => r.conditionId === cond.id);
-              const imageColorValue = overriddenCondition?.imageColor || cond.imageColor;
-              if (!rule.propertyName || !cond.value || !imageColorValue) return;
-              const stringCondition = conditionWithOperation(
-                variable(rule.propertyName),
-                string(cond.value),
-                cond.operation,
-              );
-              const numberCondition = !isNaN(Number(cond.value))
-                ? conditionWithOperation(
-                    defaultConditionalNumber(rule.propertyName),
-                    number(Number(cond.value)),
-                    cond.operation,
-                  )
-                : undefined;
-              return rule.propertyName && cond.value && imageColorValue
-                ? ([
-                    numberCondition ? `${numberCondition} || ${stringCondition}` : stringCondition,
-                    `color("${imageColorValue}")`,
-                  ] as [string, string])
-                : undefined;
-            });
-          }) ?? []
-        ).filter(isNotNullish),
-      ],
-    },
-  };
-};
-
-export const makeLabelTextExpression = (
-  comp: Component<typeof POINT_USE_LABEL_FIELD> | undefined,
-): ExpressionContainer | string | undefined => {
-  if (!comp) return;
-  const textExpression = comp.preset?.textExpression;
-  if (!textExpression?.startsWith("=")) return textExpression;
-  return {
-    expression: {
-      conditions: [["true", textExpression.substring(1)]],
-    },
-  };
-};
 
 export const useEvaluateGeneralAppearance = ({
   componentAtoms,
@@ -573,9 +188,10 @@ export const useEvaluateGeneralAppearance = ({
         marker: {
           style: pointStyle?.preset?.style,
           pointColor:
-            makeSimpleValue(pointColor) ??
-            makeConditionalExpression(pointFillColorCondition) ??
-            makeGradientExpression(pointFillGradientColor),
+            makeSimpleColorValue(pointColor, opacity?.value) ??
+            makeConditionalExpression(pointFillColorCondition, opacity?.value) ??
+            makeGradientExpression(pointFillGradientColor, opacity?.value) ??
+            makeSimpleColorWithOpacity(opacity, DEFAULT_COLOR),
           pointSize: pointSize?.preset?.defaultValue,
           pointOutlineColor: pointStroke?.preset?.color,
           pointOutlineWidth: pointStroke?.preset?.width,
@@ -583,8 +199,8 @@ export const useEvaluateGeneralAppearance = ({
             pointImageValue?.preset?.imageURL ??
             makeConditionalImageExpression(pointImageCondition),
           imageColor:
-            pointImageValue?.preset?.imageColor ??
-            makeConditionalImageColorExpression(pointImageCondition),
+            makeSimpleColorWithOpacity(opacity, pointImageValue?.preset?.imageColor) ??
+            makeConditionalImageColorExpression(pointImageCondition, opacity?.value),
           imageSize: pointImageSize?.preset?.defaultValue,
           imageSizeInMeters: pointImageSize?.preset?.enableSizeInMeters,
           show:
@@ -594,17 +210,22 @@ export const useEvaluateGeneralAppearance = ({
           labelText: makeLabelTextExpression(pointLabel),
           labelTypography: {
             fontSize: pointLabel?.preset?.fontSize,
-            color: pointLabel?.preset?.fontColor,
+            color: makeSimpleColorWithOpacity(opacity, pointLabel?.preset?.fontColor),
           },
           labelBackground: pointLabel?.preset?.background,
-          labelBackgroundColor: pointLabel?.preset?.backgroundColor,
+          labelBackgroundColor: makeSimpleColorWithOpacity(
+            opacity,
+            pointLabel?.preset?.backgroundColor,
+          ),
           height: pointLabel?.preset?.height,
           extrude: pointLabel?.preset?.extruded,
           heightReference: pointHeightReference?.preset?.defaultValue,
         },
         polyline: {
           strokeColor:
-            makeSimpleValue(polylineColor) ?? makeConditionalExpression(polylineFillColorCondition),
+            makeSimpleColorValue(polylineColor, opacity?.value) ??
+            makeConditionalExpression(polylineFillColorCondition, opacity?.value) ??
+            makeSimpleColorWithOpacity(opacity, DEFAULT_COLOR),
           strokeWidth: polylineStrokeWeight?.preset?.defaultValue,
           show:
             makeVisibilityFilterExpression(polylineVisibilityFilter) ??
@@ -616,11 +237,16 @@ export const useEvaluateGeneralAppearance = ({
         },
         polygon: {
           fillColor:
-            makeSimpleValue(polygonFillAndStrokeColor) ??
-            makeConditionalExpression(polygonFillAndStrokeColorCondition),
+            makeSimpleColorValue(polygonFillAndStrokeColor, opacity?.value) ??
+            makeConditionalExpression(polygonFillAndStrokeColorCondition, opacity?.value) ??
+            makeSimpleColorWithOpacity(opacity, DEFAULT_COLOR),
           strokeColor:
-            makeSimpleValueForStrokeColor(polygonFillAndStrokeColor) ??
-            makeStrokeColorConditionalExpression(polygonFillAndStrokeColorCondition),
+            makeSimpleValueForStrokeColor(polygonFillAndStrokeColor, opacity?.value) ??
+            makeStrokeColorConditionalExpression(
+              polygonFillAndStrokeColorCondition,
+              opacity?.value,
+            ) ??
+            makeSimpleColorWithOpacity(opacity, DEFAULT_COLOR),
           strokeWidth: polygonStrokeWeight?.preset?.defaultValue,
           stroke: !!polygonFillAndStrokeColor?.preset?.strokeValue || !!polygonStrokeWeight,
           show:
@@ -640,7 +266,7 @@ export const useEvaluateGeneralAppearance = ({
           color:
             makeConditionalExpression(tilesetFillColorCondition, opacity?.value) ??
             makeGradientExpression(tilesetFillGradientColor, opacity?.value) ??
-            makeSimpleColorWithOpacity(opacity),
+            makeSimpleColorWithOpacity(opacity, DEFAULT_COLOR),
           experimental_clipping: clippingBox,
           showWireframe: tilesetWireframe?.value?.wireframe,
         },
